requirements.txt
pyyaml>=6.0
pillow>=9.0.0
requests>=2.27.1
